- restructure filmlist a bit
